Fix a GC bug: was depending on undefined evaluation order.

+++ lst4/source/globs.h
@@ -8,7 +8,7 @@
 #include <stdio.h>
 
 extern int fileIn(FILE * fp), fileOut(FILE * fp);
-extern void sysError(char *, unsigned int), flushCache(void);
+extern void sysError(char *, unsigned long), flushCache(void);
 extern struct object *primitive(int, struct object *, int *);
 
 #endif /* GLOBS_H */
